# specs/001-package-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Elsa Package Catalog API
  version: 0.1.0
  description: Public and admin REST contracts for package, feature, compatibility, sync, validation, and approval workflows.
servers:
  - url: http://localhost:5000
security: []
...
paths:
...
  /api/admin/sync:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for all enabled sources
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
  /api/admin/sync/sources/{sourceId}:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for one source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
  /api/admin/sync/packages/{packageId}:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for one package ID
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
  /api/admin/sync-runs:
    get:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: List sync runs
      responses:
        '200':
          description: Sync runs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncRunSummary'
  /api/admin/sync-runs/{id}:
    get:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Get sync run details
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Sync run details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRunDetails'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    PackageIdPath:
      name: packageId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    SyncAccepted:
      description: Sync accepted.
      content:
        application/json:
          schema:
            type: object
            properties:
              syncRunId:
                type: string
              status:
                type: string
                enum: [Running]
...
  schemas:
...
    SyncRequest:
      type: object
      properties:
        forceReindex:
          type: boolean
          default: false
    SyncRunSummary:
      type: object
      properties:
        id:
          type: string
        trigger:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        summaryCounters:
          type: object
          additionalProperties:
            type: integer
    SyncRunDetails:
      allOf:
        - $ref: '#/components/schemas/SyncRunSummary'
        - type: object
          properties:
            error:
              type: string
            items:
              type: array
              items:
                type: object
                properties:
                  sourceId:
                    type: string
                  packageId:
                    type: string
                  version:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
                  error:
                    type: string
...